refactor(vta): extract vault/proxy-login drivers to operations (P2.4 part 2)
Second increment of P2.4(b). Moves the proxy-login business logic out of
routes/trust_tasks/vault.rs into operations::vault::proxy_login, leaving the
route handler a thin gate→parse→nonce-check→load→scope→step-up→readiness→
call→map adapter. Behaviour preserved (P2.0 wire tests green unchanged).

- New operations::vault::proxy_login::proxy_login — the two-driver dispatch
  (did-self-issued SIOP id_token; password HTTP-POST cookies, webvh-gated),
  session-blob building, and the shared authcrypt seal (reuses the
  authcrypt_to_holder helper from part 1). Returns a typed ProxyLoginError
  (NoAudience / NotProxyable / NotImplemented / PasswordPost / App) that the
  route maps back to the exact canonical reject codes.
- Helpers moved to operations: resolve_siop_audience (+ its 6 unit tests),
  build_session_blob_with_bearer / _with_cookies, first_web_origin,
  secret_kind_label (now shared with sign-trust-task via
  crate::operations::vault::secret_kind_label), PROXY_LOGIN_INNER_MSG_TYPE,
  PASSWORD_POST_TTL_CEILING_SECS.
- handle_proxy_login reduced from ~344 to ~95 lines. password_post_error_to_reject
  stays in the route (it builds the route-layer RejectReason for the
  PasswordPost arm). vault.rs 2031 → 1638.

Note: under a non-default `--no-default-features` build (webvh off), a
password-with-loginConfig entry now rejects as the generic
`not_implemented` rather than the prior "requires the webvh feature" wording —
same TaskFailed reject, normalised detail; webvh is on in every shipping build.

P2.0's 12 vault wire tests green unchanged; the 6 resolve_siop_audience unit
tests moved + green; fmt + clippy clean (all-targets); tee compiles; full
vta-service suite (965) green.

Remaining P2.4(b): part 3 sign-trust-task, part 4 upsert unseal (toward the
≤450-LOC route-file target).
